Clean up g2o example
