+++ examples/g2o.rs
@@ -3,36 +3,92 @@ use std::{env, time::Instant};
 use factrs::{
     optimizers::{GaussNewton, GraphOptimizer, Optimizer},
     utils::load_g20,
+    variables::*,
 };
-fn main() {
+
+#[cfg(feature = "rerun")]
+use factrs::rerun::RerunObserver;
+#[cfg(feature = "rerun")]
+use rerun::{Arrows2D, Arrows3D, Points2D, Points3D};
+#[cfg(feature = "rerun")]
+use std::net::{SocketAddr, SocketAddrV4};
+
+// Setups rerun and a callback for iteratively sending to rerun
+// Must run with --features rerun for it to work
+#[cfg(feature = "rerun")]
+fn rerun_init(opt: &mut GaussNewton, dim: &str, obj: &str) {
+    // Setup the rerun & the callback
+    let socket = SocketAddrV4::new("172.31.65.81".parse().unwrap(), 9876);
+    let rec = rerun::RecordingStreamBuilder::new("rerun_example_dna_abacus")
+        .connect_tcp_opts(SocketAddr::V4(socket), rerun::default_flush_timeout())
+        .unwrap();
+
+    let topic = "base/solution";
+
+    match (dim, obj) {
+        ("se2", "points") => {
+            let callback = RerunObserver::<SE2, Points2D>::new(rec, topic);
+            opt.observers.add(callback)
+        }
+        ("se2", "arrows") => {
+            let callback = RerunObserver::<SE2, Arrows2D>::new(rec, topic);
+            opt.observers.add(callback)
+        }
+        ("se3", "points") => {
+            let callback = RerunObserver::<SE3, Points3D>::new(rec, topic);
+            opt.observers.add(callback)
+        }
+        ("se3", "arrows") => {
+            let callback = RerunObserver::<SE3, Arrows3D>::new(rec, topic);
+            opt.observers.add(callback)
+        }
+        _ => panic!("Invalid arguments"),
+    };
+}
+
+#[cfg(not(feature = "rerun"))]
+fn rerun_init(_opt: &mut GaussNewton, _dim: &str, _obj: &str) {}
+
+fn main() -> Result<(), Box<dyn std::error::Error>> {
+    // ---------------------- Parse Arguments & Load data ---------------------- //
     let args: Vec<String> = env::args().collect();
     if args.len() < 2 {
-        println!("Usage: {} <g2o file>", args[0]);
-        return;
+        println!("Usage: {} <g2o file> <points/arrows>", args[0]);
+        return Ok(());
     }
-    let filename = &args[1];
 
     pretty_env_logger::init();
 
     // Load the graph from the g2o file
+    let filename = &args[1];
     let (graph, init) = load_g20(filename);
-    let to_solve = init.clone();
     println!("File loaded");
 
-    // Optimize with GaussNewton
+    // Make optimizer
     let mut optimizer: GaussNewton = GaussNewton::new(graph);
+
+    // Connect to rerun if it's been enabled
+    if cfg!(feature = "rerun") {
+        let obj = if args.len() >= 3 { &args[2] } else { "arrows" };
+        let dim = if init.filter::<SE2>().count() != 0 {
+            "se2"
+        } else if init.filter::<SE3>().count() != 0 {
+            "se3"
+        } else {
+            panic!("Graph doesn't have SE2 or SE3 variables");
+        };
+        rerun_init(&mut optimizer, dim, obj);
+    }
+
+    // ------------------------- Optimize ------------------------- //
     let start = Instant::now();
-    let result = optimizer.optimize(to_solve);
+    let result = optimizer.optimize(init);
     let duration = start.elapsed();
 
-    println!("Optimization took: {:?}", duration);
-
     match result {
-        Ok(_sol) => {
-            println!("Optimization successful!");
-        }
-        Err(e) => {
-            println!("Optimization failed: {:?}", e);
-        }
+        Ok(_) => println!("Optimization converged!"),
+        Err(e) => println!("Optimization failed: {:?}", e),
     }
+    println!("Optimization took: {:?}", duration);
+    Ok(())
 }
